chore: refactoring
tdrl parsing out of main package
